Fixed a problem where cards won't display
There was a problem where the cards won't display. I added a delay to
the state change which allowed enough time to get the fetch data and
update the component correctly. Now it loads cards after some time using
setTimeout().

Because of this, there was also a problem with the inital unit test that
also requires a delay. I am sure there is a better way in doing this but
for now it's the fix that is needed to progress.